test: LevyRndの乱数テストをクラスに分離
